project/api/requirements.txt
flask==1.1.2
psycopg2-binary
sqlalchemy==1.3.20
asyncio-nats-client
asyncio-nats-streaming